Increase file size limits

+++ ModernUpnpx/api/MediaServer1ItemRes.h
@@ -50,7 +50,7 @@
     NSString *duration;
     int nrAudioChannels;
     NSString *protocolInfo;
-    int size;
+    long long size;
     int durationInSeconds;
 }
 
@@ -58,7 +58,7 @@
 @property(retain, nonatomic) NSString *duration;
 @property(readwrite) int nrAudioChannels;
 @property(retain, nonatomic) NSString *protocolInfo;
-@property(readwrite) int size;
+@property(readwrite) long long size;
 @property(readwrite) int durationInSeconds;
 
 @end
